requirements.txt
pydirectinput>=1.0.4
pyautogui>=0.9.54
pynput>=1.7.6
keyboard>=0.13.5
mouse>=0.7.1
Pillow>=10.0.0
pywin32>=306
pyyaml>=6.0
